requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
python-dotenv
streamlit
requests
plotly
watchdog